feat(Club): 클럽 리스트 검색기능 추가 ClubRepositorySupport 인터페이스 추가로 인한 수정
issue117
